Soap4 0 headers (#482)
* Soap 4.0 API upgrades and Header changes

* 1. Changed Session version
2. Changed how we are fetching PNR from pnr reply
3. Changed a few other API version accross different files

* 1. Updates many other java files

* 1. Created Soap header

* 1. Amadeus wsdl URL's fixed
2. Changed to new creds

* 1. Action Value and Office ID login

* 1. Soap 4.0 header creating done now getting responses

* soap 4.0 header issue

* 1. Amadeus wsdl URL's fixed
2. Changed to new creds

* 1. Action Value and Office ID login

* 1. Soap 4.0 header creating done now getting responses

* Amadeus Normal Service handler Stateful and Stateless development complete

* Refund and Reissue handled

* Conflicts resolved

* Fixed Book and pay issue for Soap 4.0

* 1. Fixed Cron Logout Issue

* 1. fixed make payment error

* 1.  Fixed Reissue Session Error


* Conflicts resolved

* Ancillary fix

* Ancillary payment Wsdl port added

* ...

* Ancillary Boady mame space correction

* Added version retrive PNR (#431)

* ...

* ...

* RC entry soap 4 fix

* Updated save pnr option code from 10 to 11.

* ..

* Compilation error fixed

* ...


* 1. Soap Prod creds Specific header correction
2. Moved all creds to conf file

---------

+++ app/utils/RefundHelper.java
@@ -48,7 +48,7 @@ public class RefundHelper {
                     if(segments!=null && !segments.getSegment().isEmpty()) {
                         List<RefundedItineraryType> refundedItineraryTypeList = segments.getSegment();
                         for (RefundedItineraryType refundedSegmentTattoo : refundedItineraryTypeList) {
-                            refundedSegmentTattoos.add(refundedSegmentTattoo.getTattoo());
+                            refundedSegmentTattoos.add(BigInteger.valueOf(refundedSegmentTattoo.getTattoo()));
                         }
                     }
                     perPaxRefundPricingInformation.setRefundedSegmentTattoos(refundedSegmentTattoos);
@@ -57,7 +57,7 @@ public class RefundHelper {
                     PassengerType passengerType = passengers.getPassenger().get(0);
 
                     String paxFullName = passengerType.getFullName();
-                    BigInteger paxTattoo = passengerType.getTattoo();
+                    BigInteger paxTattoo = BigInteger.valueOf(passengerType.getTattoo());
                     perPaxRefundPricingInformation.setFullName(paxFullName);
                     perPaxRefundPricingInformation.setPaxTattoo(paxTattoo);
 
@@ -71,7 +71,7 @@ public class RefundHelper {
 
                             BigDecimal amount = taxType.getAmount();
                             String currency = taxType.getCurrencyCode();
-                            BigInteger decimalPlaces = taxType.getDecimalPlaces();
+                            BigInteger decimalPlaces = BigInteger.valueOf(taxType.getDecimalPlaces());
                             String category = taxType.getCategory();
                             String isoCode = taxType.getISOCode();
 
@@ -99,7 +99,7 @@ public class RefundHelper {
                             BigDecimal percent = penalty.getPercent();
                             BigDecimal amount = penalty.getAmount();
                             String currency = penalty.getCurrencyCode();
-                            BigInteger decimalPlaces = penalty.getDecimalPlaces();
+                            BigInteger decimalPlaces = BigInteger.valueOf(penalty.getDecimalPlaces());
                             String penaltyType = penalty.getPenaltyType();
 
                             detailedPenaltyInformation.setPercent(percent);
@@ -126,7 +126,7 @@ public class RefundHelper {
                             CommissionType.CommissionPayableAmount commissionPayableAmount = commission.getCommissionPayableAmount();
                             BigDecimal amount = commissionPayableAmount.getAmount();
                             String currency = commissionPayableAmount.getCurrencyCode();
-                            BigInteger decimalPlaces = commissionPayableAmount.getDecimalPlaces();
+                            BigInteger decimalPlaces = BigInteger.valueOf(commissionPayableAmount.getDecimalPlaces());
 
                             detailedCommissionInformation.setPercent(percent);
                             detailedCommissionInformation.setAmount(amount);
@@ -150,7 +150,7 @@ public class RefundHelper {
                             String qualifier = monetaryInformation.getQualifier();
                             BigDecimal amount = monetaryInformation.getAmount();
                             String currency = monetaryInformation.getCurrencyCode();
-                            BigInteger decimalPlaces = monetaryInformation.getDecimalPlaces();
+                            BigInteger decimalPlaces = BigInteger.valueOf(monetaryInformation.getDecimalPlaces());
 
                             detailedMonetaryInformation.setQualifier(qualifier);
                             detailedMonetaryInformation.setAmount(amount);
